Delete using SimpleSocket, use NDK-based networking There's this obscure function in the NDK called `android_setsocknetwork`. This allows the app to create a socket normally in C, and bind it to the WiFi interface. This bumps min API to 23.
Also rewrote connection setup process in C, it used to be in Gallery.java with a bunch of JNI
wrappers. All gone now.

Did the same with the tester.

Also added a button to connect to the camera with USB-OTG.. will I ever learn to stop overcomplicating things?
Well I guess it will have to be finished before the next release. Shouldn't be too hard since the frontend
JNI stuff is already an abstraction layer over most of the protocol-specific stuff.


Former-commit-id: 0fa1eb9085b55653c17f896c8634c8762f430df1